First attempt at auto removed auto choices we can do yet set time and speed based on rough testing in the shop with the second bot. dashboard cleanup

+++ robot2016/src/org/usfirst/frc2016/robot2016/commands/DelaySeconds.java
@@ -52,7 +52,7 @@ public class DelaySeconds extends Command {
 
     // Make this return true when this Command no longer needs to run execute()
     protected boolean isFinished() {
-        if (waitCounter >= (m_seconds/DELAYPERCOUNT)) {
+        if (waitCounter >= (int)(m_seconds/DELAYPERCOUNT)) {
     		return true;
     	}
         return false;
